Moved font methods from HtmlValues to FontValues. Added new font methods.
